requirements.txt
jittor==1.3.1.29
numpy==1.18.2
Pillow==9.0.0
PyYAML==6.0
tensorboardX==2.4.1
tqdm==4.46.0
